rename: components 폴더의 파일 이름 변경
