+++ src/components/Header/Header.style.ts
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import CustomAppBar from '../CustomMui/CustomAppBar';
 import forwardPropOption from '@/utils/emotionForwardPropOption';
-import { HeaderColor, HeaderPosition } from '.';
+import { HeaderColor, HeaderPosition } from './Header';
 import { css, Theme } from '@emotion/react';
 import { MAX_WIDTH } from '@/constants/css';
