+++ src/pages/Search/SearchHeader/SearchHeader.tsx
@@ -2,8 +2,8 @@ import { S } from './SearchHeader.style';
 import { Link, useLocation } from 'react-router-dom';
 import CancelIcon from '@/assets/svg/cancle.svg?react';
 import CustomTextField from '@/components/CustomMui/CustomTextField';
-import GoBackButton from '@/components/GoBackButton';
-import Header from '@/components/Header';
+import GoBackButton from '@/components/GoBackButton/GoBackButton';
+import Header from '@/components/Header/Header';
 import { SearchLocationState } from '../CurrentRegionButton/CurrentRegionButton';
 
 type SearchHeaderProps = {
