+++ src/pages/TermsOfService/TermsOfServicePage.tsx
@@ -1,4 +1,4 @@
-import HeadHelmet from '@/components/HeadHelmet';
+import HeadHelmet from '@/components/HeadHelmet/HeadHelmet';
 
 const TermsOfServicePage = () => {
   return (
